requirements_macos.txt
numpy==1.24.3
opencv-python==4.10.0.84
mediapipe==0.10.21
Pillow==10.4.0
pillow-avif-plugin==1.4.6
streamlit==1.38.0
protobuf==4.25.3

tensorflow-macos==2.13.0
tensorflow-metal==1.1.0
h5py==3.8.0
scikit-learn==1.3.2
